Partial unprivileged image building implementation
Note the `ubuntu` templte for lxc-create does not support unprivileged
building, and recommends the usage of the `download` template for such
purposes - hence the switch.

Until the implementation of this feature is finished and working, I
will leave the root UID check in place to ensure nobody attempts to
use it.

With these changes (and the root UID check disabled), the LXC image is
successfully created and vagrant hacks seem to be successfully
applied, but the Vagrant box packaging step completely fails due to
file permission errors under ~/.local/share/lxc/$CONTAINER/rootfs. I'm
not yet sure what the best approach is to solve this problem.